Move OSL templates and classes to own util file
