+++ utils/util.py
@@ -33,8 +33,7 @@ import bpy
 import bpy_extras
 from bpy.app.handlers import persistent
 
-from ..properties.nodes import generate_node, AppleseedOSLScriptNode
-from . import path_util
+from . import path_util, osl_utils
 from ..logger import get_logger
 
 logger = get_logger()
@@ -56,102 +55,6 @@ def safe_unregister_class(cls):
         bpy.utils.unregister_class(cls)
     except Exception as e:
         logger.error("[appleseed] ERROR: Failed to unregister class {0}: {1}".format(cls, e))
-
-
-def read_osl_shaders():
-    '''
-    Reads parameters from OSL .oso files using the ShaderQuery function that is built
-    into the Python bindings for appleseed.  These parameters are used to create a dictionary
-    of the shader parameters that is then added to a list.  This shader list is passed
-    on to the oslnode.generate_node function.
-    '''
-
-    nodes = []
-
-    if not path_util.get_appleseed_bin_dir_path():
-        logger.warning("[appleseed] WARNING: Path to appleseed's binary directory not set: rendering and OSL features will not be available.")
-        return nodes
-
-    shader_directories = path_util.get_osl_search_paths()
-
-    q = asr.ShaderQuery()
-
-    logger.debug("[appleseed] Parsing OSL shaders...")
-
-    for shader_dir in shader_directories:
-        if os.path.isdir(shader_dir):
-            logger.debug("[appleseed] Searching {0} for OSO files...".format(shader_dir))
-            for file in os.listdir(shader_dir):
-                if file.endswith(".oso"):
-                    logger.debug("[appleseed] Reading {0}...".format(file))
-                    filename = os.path.join(shader_dir, file)
-                    q.open(filename)
-                    nodes.append(parse_shader(q, filename=filename))
-
-    logger.debug("[appleseed] OSL parsing complete.")
-
-    return nodes
-
-def parse_shader(q, filename=None):
-    d = {}
-    d['inputs'] = []
-    d['outputs'] = []
-    shader_meta = q.get_metadata()
-    if 'as_node_name' in shader_meta:
-        d['name'] = shader_meta['as_node_name']['value']
-    else:
-        d['name'] = q.get_shader_name()
-    d['filename'] = filename
-    if 'URL' in shader_meta:
-        d['url'] = shader_meta['URL']['value']
-    else:
-        d['url'] = ''
-    if 'as_category' in shader_meta:
-        d['category'] = shader_meta['as_category']['value']
-    else:
-        d['category'] = 'other'
-    num_of_params = q.get_num_params()
-    for x in range(0, num_of_params):
-        metadata = {}
-        param = q.get_param_info(x)
-        if 'metadata' in param:
-            metadata = param['metadata']
-        param_data = {}
-        param_data['name'] = param['name']
-        param_data['type'] = param['type']
-        param_data['connectable'] = True
-        param_data['hide_ui'] = param['validdefault'] is False
-        if 'default' in param:
-            param_data['default'] = param['default']
-        if 'label' in metadata:
-            param_data['label'] = metadata['label']['value']
-        if 'widget' in metadata:
-            param_data['widget'] = metadata['widget']['value']
-            if param_data['widget'] == 'null':
-                param_data['hide_ui'] = True
-        if 'page' in metadata:
-            param_data['section'] = metadata['page']['value']
-        if 'min' in metadata:
-            param_data['min'] = metadata['min']['value']
-        if 'max' in metadata:
-            param_data['max'] = metadata['max']['value']
-        if 'softmin' in metadata:
-            param_data['softmin'] = metadata['softmin']['value']
-        if 'softmax' in metadata:
-            param_data['softmax'] = metadata['softmax']['value']
-        if 'help' in metadata:
-            param_data['help'] = metadata['help']['value']
-        if 'options' in metadata:
-            param_data['options'] = metadata['options']['value'].split(" = ")[-1].replace("\"", "").split("|")
-        if 'as_blender_input_socket' in metadata:
-            param_data['connectable'] = False if metadata['as_blender_input_socket']['value'] == 0.0 else True
-
-        if param['isoutput'] is True:
-            d['outputs'].append(param_data)
-        else:
-            d['inputs'].append(param_data)
-            
-    return d
 
 
 # ------------------------------------
@@ -181,18 +84,6 @@ def realpath(path):
     return path
 
 
-def compile_osl_bytecode(compiler, script_block):
-    osl_path = bpy.path.abspath(script_block.filepath, library=script_block.library)
-    if script_block.is_in_memory or script_block.is_dirty or script_block.is_modified or not os.path.exists(osl_path):
-        source_code = script_block.as_string()
-    else:
-        code = open(osl_path, 'r')
-        source_code = code.read()
-        code.close()
-
-    return compiler.compile_buffer(source_code)
-
-
 @persistent
 def update_project(_):
     # Compile all OSL Script nodes
@@ -200,17 +91,18 @@ def update_project(_):
     compiler = asr.ShaderCompiler(stdosl_path)
     q = asr.ShaderQuery()
     for script in bpy.data.texts:
-        osl_bytecode = compile_osl_bytecode(compiler, script)
+        osl_bytecode = osl_utils.compile_osl_bytecode(compiler, script)
         if osl_bytecode is not None:
             q.open_bytecode(osl_bytecode)
 
-            node_data = parse_shader(q)
+            node_data = osl_utils.parse_shader(q)
 
-            node_name, node_category, node_classes = generate_node(node_data, AppleseedOSLScriptNode)
+            node_name, node_category, node_classes = osl_utils.generate_node(node_data, osl_utils.AppleseedOSLScriptNode)
 
             for cls in node_classes:
                 safe_register_class(cls)
-    
+
+
 # ------------------------------------
 # Scene export utilities.
 # ------------------------------------
